src/multi-platform/ecs/requirements.txt
flask>=3.0.0
boto3>=1.28.0
requests>=2.31.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-http>=1.24.0
opentelemetry-instrumentation-flask>=0.45b0
opentelemetry-instrumentation-requests>=0.45b0
opentelemetry-instrumentation-botocore>=0.45b0
kafka-python-ng>=2.0.0
aws-msk-iam-sasl-signer-python>=1.0.0
